+++ iris_grib/tests/__init__.py
@@ -1,19 +1,8 @@
-# (C) British Crown Copyright 2010 - 2018, Met Office
+# Copyright iris-grib contributors
 #
-# This file is part of iris-grib.
-#
-# iris-grib is free software: you can redistribute it and/or modify it under
-# the terms of the GNU Lesser General Public License as published by the
-# Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# iris-grib is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License
-# along with iris-grib.  If not, see <http://www.gnu.org/licenses/>.
+# This file is part of iris-grib and is released under the LGPL license.
+# See COPYING and COPYING.LESSER in the root of the repository for full
+# licensing details.
 """
 Provides testing capabilities and customisations specific to iris-grib.
 
@@ -21,9 +10,6 @@ This imports iris.tests, which requires to be done before anything else for
 plot control reasons : see documentation there.
 
 """
-from __future__ import (absolute_import, division, print_function)
-from six.moves import (filter, input, map, range, zip)  # noqa
-import six
 
 import iris.tests
 
@@ -57,7 +43,7 @@ class IrisGribTest(IrisTest):
         as a string, or sequence of strings.
 
         """
-        if not isinstance(relative_path, six.string_types):
+        if not isinstance(relative_path, str):
             relative_path = os.path.join(*relative_path)
         return os.path.abspath(os.path.join(_RESULT_PATH, relative_path))
 
@@ -103,6 +89,6 @@ class IrisGribTest(IrisTest):
         relative path as a string, or sequence of strings.
 
         """
-        if not isinstance(relative_path, six.string_types):
+        if not isinstance(relative_path, str):
             relative_path = os.path.join(*relative_path)
         return os.path.abspath(os.path.join(_TESTDATA_PATH, relative_path))