+++ iris_grib/tests/unit/load_convert/test_grid_definition_template_40.py
@@ -1,27 +1,13 @@
-# (C) British Crown Copyright 2015 - 2017, Met Office
+# Copyright iris-grib contributors
 #
-# This file is part of iris-grib.
-#
-# iris-grib is free software: you can redistribute it and/or modify it under
-# the terms of the GNU Lesser General Public License as published by the
-# Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# iris-grib is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License
-# along with iris-grib.  If not, see <http://www.gnu.org/licenses/>.
+# This file is part of iris-grib and is released under the LGPL license.
+# See COPYING and COPYING.LESSER in the root of the repository for full
+# licensing details.
 """
 Unit tests for
 :func:`iris_grib._load_convert.grid_definition_template_40`.
 
 """
-
-from __future__ import (absolute_import, division, print_function)
-from six.moves import (filter, input, map, range, zip)  # noqa
 
 # import iris_grib.tests first so that some things can be initialised
 # before importing anything else.
@@ -56,6 +42,7 @@ class Test_regular(tests.IrisGribTest):
             'scaledValueOfEarthMinorAxis': MDI,
             'iDirectionIncrement': 22500000,
             'longitudeOfFirstGridPoint': 0,
+            'resolutionAndComponentFlags': 32,
             'Ni': 16,
             'scanningMode': 0b01000000,
             'distinctLatitudes': np.array([-73.79921363, -52.81294319,