+++ iris_grib/tests/unit/load_convert/test_product_definition_template_15.py
@@ -1,19 +1,8 @@
-# (C) British Crown Copyright 2014 - 2018, Met Office
+# Copyright iris-grib contributors
 #
-# This file is part of iris-grib.
-#
-# iris-grib is free software: you can redistribute it and/or modify it under
-# the terms of the GNU Lesser General Public License as published by the
-# Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# iris-grib is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License
-# along with iris-grib.  If not, see <http://www.gnu.org/licenses/>.
+# This file is part of iris-grib and is released under the LGPL license.
+# See COPYING and COPYING.LESSER in the root of the repository for full
+# licensing details.
 """
 Test function
 :func:`iris_grib._load_convert.product_definition_template_15`.
@@ -23,14 +12,11 @@ testing for the statistical method and spatial-processing type.
 
 """
 
-from __future__ import (absolute_import, division, print_function)
-from six.moves import (filter, input, map, range, zip)  # noqa
-
 # import iris_grib.tests first so that some things can be initialised
 # before importing anything else.
 import iris_grib.tests as tests
 
-import mock
+from unittest import mock
 
 import iris.coords
 from iris.exceptions import TranslationError
@@ -105,14 +91,14 @@ class Test(LoadConvertTest):
         section = section_4_sample()
         section['statisticalProcess'] = 999
         msg = 'unsupported statistical process type \[999\]'
-        with self.assertRaisesRegexp(TranslationError, msg):
+        with self.assertRaisesRegex(TranslationError, msg):
             self._translate(section)
 
     def test_bad_spatial_processing_code(self):
         section = section_4_sample()
         section['spatialProcessing'] = 999
         msg = 'unsupported spatial processing type \[999\]'
-        with self.assertRaisesRegexp(TranslationError, msg):
+        with self.assertRaisesRegex(TranslationError, msg):
             self._translate(section)