+++ iris_grib/tests/unit/save_rules/test_product_definition_template_1.py
@@ -1,34 +1,21 @@
-# (C) British Crown Copyright 2016, Met Office
+# Copyright iris-grib contributors
 #
-# This file is part of iris-grib.
-#
-# iris-grib is free software: you can redistribute it and/or modify it under
-# the terms of the GNU Lesser General Public License as published by the
-# Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# iris-grib is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License
-# along with iris-grib.  If not, see <http://www.gnu.org/licenses/>.
+# This file is part of iris-grib and is released under the LGPL license.
+# See COPYING and COPYING.LESSER in the root of the repository for full
+# licensing details.
 """
 Unit tests for :func:`iris_grib._save_rules.product_definition_template_1`
 
 """
 
-from __future__ import (absolute_import, division, print_function)
-from six.moves import (filter, input, map, range, zip)  # noqa
-
 # Import iris_grib.tests first so that some things can be initialised before
 # importing anything else.
 import iris_grib.tests as tests
 
+from unittest import mock
+
 from cf_units import Unit
 import gribapi
-import mock
 
 from iris.coords import CellMethod, DimCoord
 import iris.tests.stock as stock
@@ -68,7 +55,7 @@ class TestRealizationIdentifier(tests.IrisGribTest):
         cube.add_aux_coord(coord, 0)
 
         msg = "'realization' coordinate with one point is required"
-        with self.assertRaisesRegexp(ValueError, msg):
+        with self.assertRaisesRegex(ValueError, msg):
             product_definition_template_1(cube, mock.sentinel.grib)